requirements.txt
alembic==1.13.2
Flask==3.0.3
Flask_Migrate==3.0.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.31
Unidecode==1.3.8
weasyprint==62.3
Werkzeug==3.0.4
WTForms==3.1.2
